"Lock after sleep" only if the clock jumps at least 1 minute.
We cannot configure a lock-time that's smaller than 1 minute, so arguably a
suspension of the machine that lasts less than 1 minute should not trigger a
lock either. Furthermore, some machines have inaccurate clocks or strange NTP
configurations and it's not that uncommon for a clock to move forward by 3
seconds. I personally own a laptop that does that and --lockaftersleep
constantly fires even though there was no suspend and I wasn't idle either.
Using a larger threshold for --lockaftersleep has the neat side-effect of
fixing that issue.